Extract IDE related functionality into own service
