+++ src/main/kotlin/org/codefreak/codefreak/graphql/api/FileApi.kt
@@ -9,7 +9,7 @@ import org.codefreak.codefreak.auth.Authority
 import org.codefreak.codefreak.auth.hasAuthority
 import org.codefreak.codefreak.graphql.BaseResolver
 import org.codefreak.codefreak.service.AnswerService
-import org.codefreak.codefreak.service.ContainerService
+import org.codefreak.codefreak.service.IdeService
 import org.codefreak.codefreak.service.file.FileContentService
 import org.codefreak.codefreak.service.file.FileService
 import org.springframework.stereotype.Component
@@ -45,7 +45,7 @@ class FileQuery : BaseResolver(), Query {
   fun answerFiles(answerId: UUID): List<FileDto> = context {
     val answer = serviceAccess.getService(AnswerService::class).findAnswer(answerId)
     val forceSaveFiles = authorization.isCurrentUser(answer.task.owner) || authorization.currentUser.hasAuthority(Authority.ROLE_ADMIN)
-    serviceAccess.getService(ContainerService::class).saveAnswerFiles(answer, forceSaveFiles)
+    serviceAccess.getService(IdeService::class).saveAnswerFiles(answer, forceSaveFiles)
     authorization.requireAuthorityIfNotCurrentUser(answer.submission.user, Authority.ROLE_TEACHER)
     val digest = serviceAccess.getService(FileService::class).getCollectionMd5Digest(answerId)
     serviceAccess.getService(FileContentService::class).getFiles(answer.id).map {
